语法材料的增加、查询、删除；

+++ src/main/java/com/webj2eedev/ieltsnote/bo/MaterialBO.java
@@ -21,7 +21,7 @@ public class MaterialBO {
         return pdo.getUid();
     }
 
-    public int deleteMaterial(int uid) {
+    public Long deleteMaterial(int uid) {
         return dao.deleteMaterial(uid);
     }
